♻️ (#38)  - 삭제 시 유저 아이디 확인 로직 추가
삭제 시 유저 아이디 확인 로직 추가

+++ src/main/java/todolist/today/today/domain/friend/application/FriendSettingService.java
@@ -54,6 +54,9 @@ public class FriendSettingService {
     }
 
     public void deleteFriendApply(String userId, String myId) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException(userId);
+        }
         FriendRelation friendId = new FriendRelation(myId, userId);
         friendApplyRepository.deleteById(friendId);
     }
